p06 Compila but SEGV

+++ p06/include/types.hpp
@@ -2,6 +2,6 @@
 
 #include <cstdint>
 
-typedef int_fast32_t i32;
+typedef float i32;
 typedef uint_fast32_t u32;
 typedef uint_fast64_t u64;
